Increase polling interval.
When a client is just a couple of ms slower than the server, it will try to catch on.
However it is then too late, as it polls every second. Then the client is once
again behind and has to adjust again.
This process continues for 4-5 seconds until it is able to catch on.

In the end, the initial loading process is very stuttering which leads to a
worse user experience.